requirements.txt
django-ldapdb
passlib
